change timesheet endpoint to schedule and remove req.body.masterId as param
