+++ Citrus/server/interfaces/Schedule.ts
@@ -1,0 +1,6 @@
+export interface Schedule {
+  masterId: number;
+  masterName: string;
+  cost: number;
+  freetimes: Array<number>;
+}
