+++ Citrus/server/routes/calendar.route.ts
@@ -4,6 +4,6 @@ import CalendarController from '@controllers/calendar.controller';
 const router = express.Router();
 
 router.post('/calendar', CalendarController.calendar);
-router.post('/calendar/timesheets', CalendarController.timesheets);
+router.post('/calendar/schedule', CalendarController.schedule);
 
 export { router as CalendarRoutes };
